backend/requirements.txt
# Core dependencies
connexion[swagger-ui]>=2.14.2
Flask==2.1.1
flask-cors
swagger-ui-bundle>=0.0.2
python_dateutil>=2.6.0
setuptools>=21.0.0

# Database & ORM
SQLAlchemy>=1.4
PyMySQL>=1.0.2
python-dotenv>=0.21.0
